refactor(mm): M6-S1 move assemble_mm_inputs to neutral mm_core
First M6 step (non-destructive prep): relocate mm_assembly from multimodal/manager/ into the neutral
mm_core layer so the understanding-staged manager package can be deleted later (M6-S5) without
breaking the in-model path. mm_core is a sanctioned srt dependency, so the in-model encode_mm_reqs
edge drops its importlib workaround (which only existed to dodge the now-removed srt->multimodal
reverse import) for a plain import. Staged embed/vit runners + the test import from the new location;
the old file is deleted (no shim -- all four importers updated). Pure-numpy module unchanged.

Verified: test_mm_assembly 5 passed on TPU pod (JAX_PLATFORMS=cpu).
